created a seperate file for create a temperature
So, over the past two weeks, I finally picked this up in hopes
of one day getting this working. Long story short, I did! The goal
was to grab temperatures from the Pi and have the Pi 'software'
send it up to this API that I have created and store it in a
database.

For the commit:
I have added a scanning system that will scan the first query for
any rows that have UUID in the JSON blob handed to the API